test:(ownership) separate shared contracts from Codex suite
