+++ test/shared-skill/memorax-code-skill.test.mjs
@@ -16,7 +16,7 @@ import { dirname, join, resolve } from "node:path";
 import { test } from "node:test";
 import { fileURLToPath } from "node:url";
 
-const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "..");
+const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "../../packages/ts/memorax-code-codex-adapter");
 const skillsRoot = join(packageRoot, "skills");
 const skillRoot = join(skillsRoot, "memorax-code");
