+++ test/shared-skill/repo-memory-builder-collect-all.test.mjs
@@ -6,7 +6,7 @@ import { dirname, join, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { test } from "node:test";
 
-const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "..");
+const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "../../packages/ts/memorax-code-codex-adapter");
 const builderSkillRoot = join(packageRoot, "skills", "memorax-code");
 const repoMemoryScript = join(builderSkillRoot, "scripts", "repo-memory.mjs");
 const defaultsPath = join(builderSkillRoot, "defaults.json");
