+++ test/shared-skill/repo-memory-builder-validate.test.mjs
@@ -6,7 +6,7 @@ import { dirname, join, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { test } from "node:test";
 
-const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "..");
+const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "../../packages/ts/memorax-code-codex-adapter");
 const repoMemoryScript = join(packageRoot, "skills", "memorax-code", "scripts", "repo-memory.mjs");
 
 function writeJson(path, value) {
