+++ test/shared-skill/repo-memory-reader.test.mjs
@@ -4,7 +4,7 @@ import { dirname, join, resolve } from "node:path";
 import { test } from "node:test";
 import { fileURLToPath } from "node:url";
 
-const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "..");
+const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "../../packages/ts/memorax-code-codex-adapter");
 const readerSkillRoot = join(packageRoot, "skills", "memorax-code");
 
 test("memorax-code repo-read reference enforces retrieval budget and stop rules", () => {
