+++ test/shared-skill/repo-memory-updater.test.mjs
@@ -6,7 +6,7 @@ import { dirname, join, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { test } from "node:test";
 
-const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "..");
+const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "../../packages/ts/memorax-code-codex-adapter");
 const skillRoot = join(packageRoot, "skills", "memorax-code");
 const repoMemoryScript = join(skillRoot, "scripts", "repo-memory.mjs");
