+++ test/shared-skill/repo-procedure-memory-skill.test.mjs
@@ -4,7 +4,7 @@ import { dirname, join, resolve } from "node:path";
 import { test } from "node:test";
 import { fileURLToPath } from "node:url";
 
-const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "..");
+const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), "../../packages/ts/memorax-code-codex-adapter");
 const skillRoot = join(packageRoot, "skills", "memorax-code");
 
 function readSkill(path) {
